renamed day_button.js into day_night_button.js, also changed path to file; changed day_night_button.js and _button.scss, now day/night will be tracking by file day_night.js, styles also changed for working with this configuration; day_night_button.js working now only for switch button
