+++ src/index.js
@@ -1,7 +1,8 @@
 import './JS/api_key';
 import { openModalOops } from './JS/modal-oops';
 import { openModalPopUp } from './JS/modal-pop-up';
-import './JS/day_button';
+import './JS/day_night_button';
+import './JS/day_night';
 import './JS/modal-menu';
 import './JS/hero';
 import './JS/teamModal';
