+++ src/sass/_button.scss
@@ -1,7 +1,3 @@
-// button { 
-//     margin: 10px; //! delete before push 
-// }
-
 //! Стилі для мобайл
 @media screen and (min-width: 320px) {
     .load-btn,
@@ -21,10 +17,10 @@
             color: var(--black-bg);
             transition: color var(--transit);
             position: relative;
-            z-index: 1000;
+            z-index: 998;
     }
 
-    .more-details-btn  {
+    .more-details-btn {
         background-color: transparent;
         color: var(--white);
     }
@@ -34,8 +30,7 @@
     .trailer-btn::after,
     .remind-btn::after, 
     .get-started-btn::after,
-    .orange-add-btn::after
-    { 
+    .orange-add-btn::after { 
         content: "";
         position: absolute;
         top: 0px;
@@ -48,14 +43,12 @@
         z-index: -1;
     }
 
-
     .load-btn::before,
     .search-btn::before,
     .trailer-btn::before,
     .remind-btn::before,
     .get-started-btn::before,
-    .orange-add-btn::before
-    {
+    .orange-add-btn::before {
         content: "";
         position: absolute;
         top: -1px;
@@ -88,14 +81,9 @@
         color: var(--white);
         transition: color var(--transit);
         position: relative;
-        z-index: 1000;
+        z-index: 998;
     }   
-    
-    .remove-btn.day, .add-btn.day {
-        color: var(--black-bg);
-        transition: color var(--transit);
-    }    
-
+  
     .remove-btn::after, .add-btn::after {
         content: "";
         position: absolute;
@@ -109,11 +97,6 @@
         z-index: -1;
     }
 
-    .remove-btn.day::after, .add-btn.day::after {
-        background-color: var(--white);
-        transition: background-color var(--transit);
-    }
-
     .remove-btn::before, .add-btn::before {
         content: "";
         position: absolute;
@@ -125,13 +108,12 @@
         transition: background var(--transit);
         border-radius: 74px;
         z-index: -2;
-        }
+    }
 
     .remove-btn:hover, 
     .remove-btn:focus, 
     .add-btn:hover,
-    .add-btn:focus 
-     { 
+    .add-btn:focus { 
         color: var(--accent);
     }
 
@@ -160,19 +142,6 @@
 
     .more-details-btn:hover, .more-details-btn:focus { 
         color: var(--accent);
-    }
-
- 
-
-    .remove-btn.day:hover::after,
-    .remove-btn.day:focus::after,
-    .add-btn.day:hover::after,
-    .add-btn.day:focus::after,
-    .remove-btn.day:hover::before,
-    .remove-btn.day:focus::before,
-    .add-btn.day:hover::before,
-    .add-btn.day:focus::before {
-        background: var(--black-bg);
     }
 
     .switch {
@@ -227,21 +196,6 @@
         visibility: hidden;
     }
 
-    .switch.day::after {
-        right: 0px; 
-        transform: translateX(-29px);
-    }
-
-    .switch.day .icon-moon-line {
-        right: 0px;
-        transform: translate(-38px, 0px) rotateY(180deg);
-    }
-
-    .switch.day .icon-moon {
-        transform: rotate(360deg);
-        z-index: -1;
-    }
-
     .icon-sun {
         position: absolute;
         width: 9.2px;
@@ -250,11 +204,6 @@
         top: 4.5px;
         z-index: -1;
         transition: transform var(--transit);
-    }
-
-    .switch.day .icon-sun {
-        z-index: 2;
-        transform: rotate(45deg);
     }
 
     .global-search-btn {
@@ -315,15 +264,74 @@
         stroke: var(--accent);
     }
 
-    .global-search-btn.day .search-icon {
-        stroke: var(--black-bg);
+    body.is-light-mode {
+        .remove-btn, .add-btn {
+            color: var(--black-bg);
+        }   
+        
+        .remove-btn:hover, .add-btn:hover,
+        .remove-btn:focus, .add-btn:focus {
+            color: var(--accent);
+        }    
+
+        .remove-btn::after, .add-btn::after {
+            background-color: var(--white);
+            transition: background-color var(--transit);
+        }
+
+        .remove-btn:hover::after,
+        .remove-btn:focus::after,
+        .add-btn:hover::after,
+        .add-btn:focus::after,
+        .remove-btn:hover::before,
+        .remove-btn:focus::before,
+        .add-btn:hover::before,
+        .add-btn:focus::before {
+            background: var(--black-bg);
+        }
+            
+        .switch::after {
+            right: 0px; 
+            transform: translateX(-29px);
+        }
+
+        .switch .icon-moon-line {
+            right: 0px;
+            transform: translate(-38px, 0px) rotateY(180deg);
+        }
+
+        .switch .icon-moon {
+            transform: rotate(360deg);
+            z-index: -1;
+        }
+        
+        .switch .icon-sun {
+            z-index: 2;
+            transform: rotate(45deg);
+        }
+
+        .global-search-btn .search-icon {
+            stroke: var(--black-bg);
+        }
+        
+        .global-search-btn:hover .search-icon, 
+        .global-search-btn:focus .search-icon  { 
+            stroke: var(--accent);
+        }
+
+        .switch::after {
+            transform: translateX(-39px);
+        }
+      
+        .switch .icon-moon-line {
+            transform: translate(-52px, 0px) rotateY(180deg);
+        }
+      
+        .switch .icon-moon {
+            transform: rotate(360deg);
+            z-index: -1;
+        }
     }
-    
-    .global-search-btn.day:hover .search-icon, 
-    .global-search-btn.day:focus .search-icon  { 
-        stroke: var(--accent);
-    }
-   
 }
 //! Стилі для таблета
 @media screen and (min-width: 768px) {
@@ -369,18 +377,7 @@
       transform: translate(50px, 0);
      }
 
-    .switch.day::after {
-      transform: translateX(-39px);
-     }
-
-     .switch.day .icon-moon-line {
-      transform: translate(-52px, 0px) rotateY(180deg);
-    }
-
-     .switch.day .icon-moon {
-      transform: rotate(360deg);
-      z-index: -1;
-     }
+   
 
     .icon-sun {
         width: 14px;
